Update to DC-DOS 3.1
Includes an IMAGE command that can tolerate empty sectors (with a sector
count) and actually working versions of Moon Patrol and DCPU-MUD. I
tested them on Lettuce, so they definitely work.

Add Fen-Ok's Pixelator to DC-DOS

It will clobber the filesystem if you try to save to the DC-DOS disk,
but it appears to load fine off disk! Now you can art!